+++ R/getTrend.R
@@ -3,8 +3,8 @@
 #' @description
 #' This function uses [terra::app()] to apply the [Kendall::MannKendall()] function to each pixel of a multi-layer time series SpatRaster, testing for a monotonic (either increasing or decreasing) trend in the raster values.
 
-#' @param rasts multi-layer SpatRaster with the output of [getPredictions()], or another time series of values for which to detect a trend
-#' @param occs species occurrence coordinates (2 columns in this order: x, y or LONGitude, LATitude) in an object coercible to a data.frame (e.g. a data.frame, matrix, tibble, sf object or SpatVector of points), and in the same coordinate reference system as 'rasts'. If provided, output pixels that do not overlap these points will be NA
+#' @param rasts multi-layer SpatRaster with the output of [getPredictions()], or another time series of values for which to detect a trend. Note more than 3 non-NA values are required for a trend to be computed.
+#' @param occs SpatVector of species occurrence points, or their spatial coordinates (2 columns in this order: x, y or LONGitude, LATitude) in an object coercible to a data.frame (e.g. a data.frame, matrix, tibble, sf object), and in the same coordinate reference system as 'rasts'. If provided, output pixels that do not overlap these points will be NA
 #' @param alpha numeric value indicating the threshold significance level for Kendall's tau statistic. Default 0.05. Pixels with p-value above this will be NA in the output.
 #' @param full logical value indicating whether to output a multi-layer raster with the full results of the Mann-Kendall test, namely the Tau value, p-value (significance), S and variance of S. If set to FALSE, a single-layer raster will be returned with the (significant) Tau values.
 #' @param file optional file name (including path, not including extension) if you want the outputs raster(s) to be saved on disk. If 'file' already exists in the working directory, the rasters are imported from there.
@@ -15,7 +15,7 @@
 #' @author A. Marcia Barbosa
 #' @seealso [Kendall::MannKendall()], spatialEco::raster.kendall()
 #' @export
-#' @importFrom terra app ifel mask rast vect writeRaster
+#' @importFrom terra app crs ifel mask project rast vect writeRaster
 #' @importFrom Kendall MannKendall
 #'
 #' @examples
@@ -40,10 +40,12 @@ getTrend <- function(rasts, occs = NULL, alpha = 0.05, file = NULL, full = TRUE,
       return(unlist(Kendall::MannKendall(x)))
     } else {
       return(c(tau = NA, sl = NA, S = NA, D = NA, varS = NA))
-    }
+    } # and see warning below
   }
 
   out <- terra::app(rasts, mannkend)
+
+  if (all(!is.finite(values(out))))  warning("Insufficient data to assess trend.")
 
   out <- out[[-grep("D", names(out))]]  # denominator, tau=S/D (as per ?Kendall::MannKendall)
 
@@ -56,10 +58,24 @@ getTrend <- function(rasts, occs = NULL, alpha = 0.05, file = NULL, full = TRUE,
 
   # remove non-occurrence pixels:
   if (!is.null(occs)) {
-    if (inherits(occs, "data.frame")) {
-      if (ncol(occs) > 2) message ("assuming 'occs' coordinates are in the first two columns")
-      occs <- terra::vect(occs, geom = colnames(occs)[1:2])
+
+    if (inherits(occs, "sf")) {
+      occs <- terra::vect(occs)
     }
+
+    if (inherits(occs, "data.frame")) {
+      if (verbosity > 0 && ncol(occs) > 2)
+        message ("assuming 'occs' X-Y coordinates are in the first two columns,\nand in the same CRS as 'rasts'")
+      occs <- terra::vect(occs,
+                          geom = colnames(occs)[1:2],
+                          crs = terra::crs(rasts))
+    }
+
+    if (inherits(occs, "SpatVector") && !terra::same.crs(occs, out)) {
+      if (verbosity > 1) message ("projecting 'occs' to the same CRS as 'rasts'")
+      occs <- terra::project(occs, terra::crs(rasts))
+    }
+
     out <- terra::mask(out, occs)
   }